SuiteCRM 7.4 Sprint Two / Amend Products Purchased Sub-Panel / QTY – Change Format to show no decimals if a decimal is not provided as the quantity.

+++ modules/AOS_Products_Quotes/vardefs.php
@@ -112,7 +112,7 @@ $dictionary['AOS_Products_Quotes'] = array(
     'required' => false,
     'name' => 'product_qty',
     'vname' => 'LBL_PRODUCT_QTY',
-    'type' => 'decimal',
+    'type' => 'integer',  // it was: 'type' => 'decimal',
     'massupdate' => 0,
     'comments' => '',
     'help' => '',
